requirements.txt
acvl_utils==0.2.2
anyio==4.8.0
asttokens==3.0.0
batchgenerators==0.25.1
batchgeneratorsv2==0.2.1
blosc2==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
connected-components-3d==3.22.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dicom2nifti==2.5.1
dynamic_network_architectures==0.3.1
einops==0.8.0
executing==2.1.0
fft-conv-pytorch==1.2.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
future==1.0.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagecodecs==2024.12.30
imageio==2.36.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
ndindex==1.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nnunetv2==2.5.1
numexpr==2.10.2
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydicom==3.0.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gdcm==3.0.24.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
SimpleITK==2.4.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.2
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
unittest2==1.1.0
urllib3==2.3.0
wcwidth==0.2.13
yacs==0.1.8
